Refactoring de la seccion account
